Add search translation keys and update latest search count constant
