+++ lib/product/model/constant/project_general_constant.dart
@@ -18,4 +18,6 @@ final class ProjectGeneralConstant {
 
   /// String constant
   static const String phoneNumberHint = '+90 5XX XXX XX XX';
+
+  static const int maxLatestSearchCount = 5;
 }
